requirements.txt
streamlit==1.28.0
plotly==5.15.0
pandas==2.0.3
cassandra-driver==3.28.0
kafka-python==2.0.2
requests==2.31.0
python-binance==1.0.19
python-dotenv==1.0.0
psycopg2-binary==2.9.7
